+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_pl.dart
@@ -2179,6 +2179,10 @@ class DartPdfEditorLocalizationsPl extends DartPdfEditorLocalizations {
   String get splitFailed => 'Nie udało się podzielić tego pliku PDF.';
 
   @override
+  String get guidesSnapHint =>
+      'Przyciągaj krawędzie i środki adnotacji • Przytrzymaj Alt, aby pominąć';
+
+  @override
   String get tbToolContentDelete => 'Wymaż';
 
   @override